+++ src/index.d.ts
@@ -1,4 +1,6 @@
 export * from './Spinorama/index.types'
 export * from './Wrapper/index.types'
 export * from './Item/index.types'
+export * from './Actions/index.types'
+export * from './Thumbnails/index.types'
 export * from './Buttons/index.types'