+++ src/index.ts
@@ -1,4 +1,6 @@
 export { default as Spinorama } from './Spinorama'
 export { default as SpinoramaWrapper } from './Wrapper'
 export { default as SpinoramaItem } from './Item'
+export { default as SpinoramaActions } from './Actions'
+export { default as SpinoramaThumbnails } from './Thumbnails'
 export { default as SpinoramaButtons } from './Buttons'